+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: yallo <yallo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/14 12:34:32 by yallo             #+#    #+#             */
-/*   Updated: 2022/11/15 13:57:14 by yallo            ###   ########.fr       */
+/*   Updated: 2022/11/16 13:40:41 by yallo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dst, const void *src, size_t size)
 {
-	int		i;
+	size_t	i;
 	char	*buf;
 	char	*temp;
